+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project changes will be shown here.
 ### Changed
 - Issue 137 - CNC029 - Password archiving & security
 - Issue 436 - Changing contracts in SR, permissions required
+### Added
+- Issue 442 - Customer Starter & Leaver questions
 
 ## [v1.16.1] - 2019-01-29
 ### Fixed